backend/requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
asyncpg
python-dotenv
google-generativeai
PyPDF2
python-multipart
pydantic
razorpay
bcrypt
httpx
